BUG - Hot Keys throw exceptions. This could be related to an invalid path.
ADD - Change all DEBUG event handlers to validate the source of the test
data folders --AND-- validate the source and target folders.
ADD - Add message box that notifies user that a validation failed.
Pressing OK cancels the DEBUG Hot-Key.

+++ Push/MainForm.cs
@@ -330,10 +330,21 @@ namespace Push
 
 		private void DEBUG_MistyRose()
 		{
-			DEBUG_InitFolders();
+			string SourceTestData = @"C:\DEV_TESTDATA\Pictures";
+			string TargetTestData = @"C:\DEV_TESTDATA\TargetPictures";
+		
+			if (DEBUG_InitFolders())
+				return;
 
-			DEBUG_LoadFolderTestData(@"C:\DEV_TESTDATA\Pictures", appSettings.SourcePath);
-			DEBUG_LoadFolderTestData(@"C:\DEV_TESTDATA\TargetPictures", appSettings.TargetPath);
+			// Validate the test data folders... 
+			if (!Directory.Exists(SourceTestData) || !Directory.Exists(TargetTestData))
+			{
+				MessageBox.Show("The Debug test data is not available.\nDEBUG Hot-Key Canceled");
+				return;
+			}
+
+			DEBUG_LoadFolderTestData(SourceTestData, appSettings.SourcePath);
+			DEBUG_LoadFolderTestData(TargetTestData, appSettings.TargetPath);
 
 			//-----------------------------------------------------------------
 			// Clear the status list box...
@@ -344,14 +355,23 @@ namespace Push
 		} // END_METHOD
 
 
-		private void DEBUG_InitFolders()
+		private bool DEBUG_InitFolders()
 		{
+			// Validate the source and target folders...
+			if (!Directory.Exists(appSettings.SourcePath) || !Directory.Exists(appSettings.TargetPath))
+			{
+				MessageBox.Show("The Source or Target Path do NOT Exist.\nDEBUG Hot-Key Canceled");
+				return true;
+			}
+			
 			// Generate a collection of ALL files, source and target, that must be deleted...
 			List<string> fileList = new List<string>(Directory.GetFiles(appSettings.SourcePath));
 			fileList.AddRange(new List<string>(Directory.GetFiles(appSettings.TargetPath)));
 			
 			foreach (string file in fileList) 
 				File.Delete(file);
+
+			return false;
 		} // END_METHOD
 
 		
@@ -375,10 +395,23 @@ namespace Push
 
 		private void DEBUG_PaleGreen()
 		{
-			DEBUG_InitFolders();
+			string SourceTestData = @"C:\DEV_TESTDATA\Pictures";
+			string TargetTestData = @"C:\DEV_TESTDATA_2";
 
-			DEBUG_LoadFolderTestData(@"C:\DEV_TESTDATA\Pictures", appSettings.SourcePath);
-			DEBUG_LoadFolderTestData(@"C:\DEV_TESTDATA_2", appSettings.TargetPath);
+			// Validate Source and Target folders...
+			if (DEBUG_InitFolders())
+				return;
+
+			// Validate the test data folders... 
+			if (!Directory.Exists(SourceTestData) || !Directory.Exists(TargetTestData))
+			{
+				MessageBox.Show("The Debug test data is not available.\nDEBUG Hot-Key Canceled");
+				return;
+			}
+
+
+			DEBUG_LoadFolderTestData(SourceTestData, appSettings.SourcePath);
+			DEBUG_LoadFolderTestData(TargetTestData, appSettings.TargetPath);
 
 			//-----------------------------------------------------------------
 			// Clear the status list box...
@@ -391,10 +424,24 @@ namespace Push
 
 		private void DEBUG_PowderBlue()
 		{
-			DEBUG_InitFolders();
+			//DEBUG_InitFolders();
+			string SourceTestData = @"C:\DEV_TESTDATA\Pictures";
+			string TargetTestData = @"C:\DEV_TESTDATA_1";
 
-			DEBUG_LoadFolderTestData(@"C:\DEV_TESTDATA\Pictures", appSettings.SourcePath);
-			DEBUG_LoadFolderTestData(@"C:\DEV_TESTDATA_1", appSettings.TargetPath);
+			// Validate Source and Target folders...
+			if (DEBUG_InitFolders())
+				return;
+
+			// Validate the test data folders... 
+			if (!Directory.Exists(SourceTestData) || !Directory.Exists(TargetTestData))
+			{
+				MessageBox.Show("The Debug test data is not available.\nDEBUG Hot-Key Canceled");
+				return;
+			}
+
+
+			DEBUG_LoadFolderTestData(SourceTestData, appSettings.SourcePath);
+			DEBUG_LoadFolderTestData(TargetTestData, appSettings.TargetPath);
 
 			//-----------------------------------------------------------------
 			// Clear the status list box...
@@ -407,10 +454,22 @@ namespace Push
 
 		private void DEBUG_Pink()
 		{
-			DEBUG_InitFolders();
+			string SourceTestData = @"C:\DEV_TESTDATA_3\Source";
+			string TargetTestData = @"C:\DEV_TESTDATA_3\Target";
 
-			DEBUG_LoadFolderTestData(@"C:\DEV_TESTDATA_3\Source", appSettings.SourcePath);
-			DEBUG_LoadFolderTestData(@"C:\DEV_TESTDATA_3\Target", appSettings.TargetPath);
+			// Validate Source and Target folders...
+			if (DEBUG_InitFolders())
+				return;
+
+			// Validate the test data folders... 
+			if (!Directory.Exists(SourceTestData) || !Directory.Exists(TargetTestData))
+			{
+				MessageBox.Show("The Debug test data is not available.\nDEBUG Hot-Key Canceled");
+				return;
+			}
+
+			DEBUG_LoadFolderTestData(SourceTestData, appSettings.SourcePath);
+			DEBUG_LoadFolderTestData(TargetTestData, appSettings.TargetPath);
 
 			//-----------------------------------------------------------------
 			// Clear the status list box...
